Modified ushort to int in Q1

+++ Answers/Question1.cs
@@ -10,13 +10,13 @@ namespace C_Sharp_Challenge_Skeleton.Answers
     {
         public static int Answer(int[] portofolios)
         {
-            ushort max = UInt16.MinValue;
+            int max = UInt16.MinValue;
 
             for (int i = 0; i < portofolios.Length; i++)
             {
                 for (int ii = i + 1; ii < portofolios.Length; ii++)
                 {
-                    ushort result = (ushort)(portofolios[i] ^ portofolios[ii]);
+                    int result =(portofolios[i] ^ portofolios[ii]);
                     if (result > max)
                     {
                         max = result;
@@ -24,7 +24,7 @@ namespace C_Sharp_Challenge_Skeleton.Answers
                 }
             }
 
-            return (int)max;
+            return max;
         }
     }
 }
